Add files via upload
the best machine learning tutorials on the web

This is a machine learning module I found here:

http://www.dcs.gla.ac.uk/~girolami/Machine_Learning_Module_2006/week_2/Lectures/wk_2_lect_2.pdf

None of this material is mine, it has all been created by Professor. M. A .Girolami.
This is hands down the best machine-learning tutorials I have found on the web, and I was afraid
the university link would be taken down, so now its on github.

i hope you enjoy this as much as i did.